requirements.txt
flask
python-dotenv
openai>=1.0.0
anthropic>=0.25.0
